fix(Loop):next try

+++ src/app/member-guard.service.ts
@@ -26,16 +26,17 @@ export class MemberGuardService implements CanActivate {
   }
 
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> | boolean {
-    let cookieValue: string = this.cookieService.get('redirect_after_login');
+    let cookieValue: string;
     this.cookieService.delete('redirect_after_login');
     if (this.cookieService.check('redirect_after_login')) {
-
+      cookieValue = this.cookieService.get('redirect_after_login');
+      this.cookieService.delete('redirect_after_login');
       this.cookieService.set('redirect_after_login', null, now());
     }
     console.log('###');
 
-    console.log(this.cookieService.get('redirect_after_login'));
-    console.log(cookieValue);
+    console.log('Cookie: ' + this.cookieService.get('redirect_after_login'));
+    console.log('CookieValue: ' + cookieValue);
     console.log('###');
 
     let redirect_url: string = state.url;
@@ -53,7 +54,7 @@ export class MemberGuardService implements CanActivate {
           return this.router.parseUrl('/registration-info');
 
         }
-        if (cookieValue) {
+        if (cookieValue && cookieValue != null && cookieValue !== 'null') {
           this.cookieService.delete('redirect_after_login');
           if (this.cookieService.check('redirect_after_login')) {
 
@@ -62,11 +63,11 @@ export class MemberGuardService implements CanActivate {
           let val: string = cookieValue;
           val = val.substring(2);
           val = val.substring(0, val.length - 1);
-          cookieValue = null;
           console.log('---');
+          console.log('Cookie: ' + this.cookieService.get('redirect_after_login'));
+          console.log('CookieValue: ' + cookieValue);
+          cookieValue = null;
 
-          console.log(this.cookieService.get('redirect_after_login'));
-          console.log(cookieValue);
           console.log('--');
 
           return this.router.parseUrl(val);
@@ -78,23 +79,4 @@ export class MemberGuardService implements CanActivate {
     }))
 
   }
-
-  canActivateChild(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean | UrlTree> | boolean {
-    this.cookieService.delete('redirect_after_login');
-
-    return this.userservice.getOnlyLoggedUserWithRedirect(null).pipe(switchMap(res => {
-      if (res['error']) {
-        window.location.href = environment.login;
-      }
-
-      return this.userservice.getMemberByUser().pipe(map(memberinfo => {
-        if (memberinfo['name'] === 'MemberNotExistsException') {
-          return this.router.parseUrl('/registration-info');
-
-        }
-        return true;
-      }))
-    }))
-  }
-
 }
